requirements.txt
numpy==1.18.5
pandas==1.3.4
tensorflow-gpu==2.3.0
tensorflow-addons==0.13.0
CUDNN==7.6.5